Upgrade to Spring Boot 2
When we upgrade our sample app to Spring Boot 2, several dependencies
will be upgraded as well.

The webjars locator feature was using the wrong dependency in Spring
Boot 1 so we've taken the opportunity in Spring Boot 2 to fix that. We
no longer provide dependency management for the former dependency so
that you notice it on upgrade.

Also, Spring Boot 2 comes with Spring Framework 5 and a Java 8 base
so we can benefit from new features such as default methods in interfaces
that makes the Adapter layer unnecessary.