Migrate properties
Datasource initialization now happens only for embedded database so the
property can be removed as it does what we want by default.

Customization of the health endpoint has a new property.

We have to remove roles configuration as the Security configuration has
changed in Spring Boot 2.

Once we're done with the migration, we can remove the dependency.