Duration support
This commit replaces our dummy delay to a `java.time.Duration`. Spring
Boot not only support the standard duration format, but also a more
readable and simple format (e.g. 3s for 3 seconds).

This is also supported by the metadata generator, making sure to detect
the unit and producing a easy to read default. For backward compatible
reason, a number can be still provided and is considered to be ms by
default. This can be customized using `DurationUnit` at the field level.